# backend/prep_frontend/openapi.yaml
openapi: 3.0.3
info:
  title: TEL-U Shop API (Backend Source)
  version: 1.0.0
servers:
  - url: http://localhost/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
    ResponseEnvelope:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
          nullable: true
        data:
          nullable: true
        errors:
          nullable: true
      additionalProperties: true
...
paths:
...
  /merchant/orders/{id}/status:
    put:
      summary: Update merchant order status
      tags: [Merchant]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
              required: [status]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope'
...
  /admin/orders/{id}/status:
    put:
      summary: Update admin order status
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
              required: [status]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseEnvelope'
...